backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.9.0
bcrypt==4.2.0
python-multipart==0.0.12
httpx==0.27.2
alembic==1.14.0
python-dateutil==2.9.0
email-validator==2.2.0
